backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
python-dotenv==1.0.0
xgboost==2.0.3
pytest==7.4.3
hypothesis==6.92.1
